Major change in KMultiAcc --- Add K-fold CV
Use K-fold CV of (X_wit_val, y_wit_val) to tune hyperparameter gamma, lambda.
In each iteration, X_wit is used to fit the witness function, X_val is used to obtain best gamma and lambda.
The optimal gamma and lambda are taken as the average of the optimal in each iteration.